Factor out resource-file access
